Update fedora-kde-min-packages.sh
Removed packages: kinfocenter, system-config-language and system-config-keyboard
Changed adwaita-gtk2-theme to breeze-gtk
Removed but recommended: firewall-config. This will be in the readme list of recommended but not necessary packages. This is a GUI for working with firewalld which is included and enabled by default in Fedora as far as I am aware. Handy for those not comfortable working with firewalld via cli.

Thanks @roesz and @PsiACE for their help on this!